Move code shared between the CLI and web server into pkg
