+++ pkg/profiles/profiles.go
@@ -1,12 +1,22 @@
 package profiles
 
 import (
+	"bytes"
 	"errors"
+	"io"
+	"text/template"
 
 	pb "github.com/linkerd/linkerd2-proxy-api/go/destination"
 	sp "github.com/linkerd/linkerd2/controller/gen/apis/serviceprofile/v1alpha1"
 	"github.com/linkerd/linkerd2/pkg/util"
 )
+
+type ProfileTemplateConfig struct {
+	ControlPlaneNamespace string
+	ServiceNamespace      string
+	ServiceName           string
+	ClusterZone           string
+}
 
 func ToRoute(route *sp.RouteSpec) (*pb.Route, error) {
 	cond, err := ToRequestMatch(route.Condition)
@@ -297,4 +307,28 @@ func ValidateResponseMatch(rspMatch *sp.ResponseMatch) error {
 	}
 
 	return nil
+}
+
+func BuildConfig(namespace, service, controlPlaneNamespace string) *ProfileTemplateConfig {
+	return &ProfileTemplateConfig{
+		ControlPlaneNamespace: controlPlaneNamespace,
+		ServiceNamespace:      namespace,
+		ServiceName:           service,
+		ClusterZone:           "svc.cluster.local",
+	}
+}
+
+func RenderProfileTemplate(config *ProfileTemplateConfig, w io.Writer) error {
+	template, err := template.New("profile").Parse(Template)
+	if err != nil {
+		return err
+	}
+	buf := &bytes.Buffer{}
+	err = template.Execute(buf, config)
+	if err != nil {
+		return err
+	}
+
+	_, err = w.Write(buf.Bytes())
+	return err
 }
